internal/proxy: add Server
A Server is added to internal/proxy, which can be used to create a local
proxy server. In addition to the existing functionality of the test proxy,

Server supports the following:
- AddModule adds an additional module to the dataset of the proxy server
- The result of the latest and list endpoints are updated dynamically if
  AddModule is called
- AddRoute adds an additional route to the proxy server

TestProxyServer now accepts a *Server instead of a *http.ServeMux. This
will be used for tests for frontend fetch in later CLs.

Change-Id: Iccc46f4c47d4b90bc1fd9bb061f2798c5dbb5b9d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/245119
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>