internal/frontend: return 404 on invalid stdlib paths
At the moment, pkg.go.dev/404 returns the following error message:

Package “404” could not be found, but you can view module “std” at
https://pkg.go.dev/mod/std.

This is because stdlib.Contains(404) returns true.

This error is fixed so that pkg.go.dev/404 just shows a normal 404
error, when frontend fetch is active.

For golang/go#37002

Change-Id: I787a10009294d105d3f92df19790dcd438123f8d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/245878
Reviewed-by: Jonathan Amsterdam <jba@google.com>