internal/frontend: enqueue tasks in serveDetails using separate context
When calling ScheduleFetch in serveDetails, it is possible that the
request ends and the context is canceled, before the goroutine finishes
enqueuing a task to the frontend fetch task queue.

A context.Background with a timeout is used instead.

Change-Id: Iee6909c1318c618e4ad3a275d281274c0be87355
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/246217
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>