content/static: properly disable details tabs
This change fixes an issue where tabs on details pages were not
properly disabled. In the case where a tab should be disabled,
the href tag is omitted from the link to take it out of the tab
order and the text is properly styled (grayed out with no hover
styling).

In the case of the overflow menu, an additional attribute
data-always-disabled is set for items that should remain disabled
regardless of whether they are shown in the tab list or not.

Fixes golang/go#40571

Change-Id: I2faeb5413493d38737c2c7fadd6f0b9a224c5e7d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/246760
Run-TryBot: Andrew Bonventre <andybons@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>