internal/database: downgrade context canceled errors to debug level
There are many places in our logs when a query will be canceled, because
all unfinished search queries for a given request are canceled.

We don't want to log these as errors, because it makes the logs very
noisy. These are downgraded to a debug level log.

Change-Id: Iaddce240394cfb4eb79b03f384e5dabbd8e84ea5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/247280
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Robert Findley <rfindley@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>