internal/postgres: move orderByLatest to path.go
This change moves orderByLatest to path.go as we transition into a path
based data model. Additionally "incompatible" is added so that we would
prefer non-incompatible modules.

Since there was a lot of duplicated logic for the ORDER BY keyword for
the queries for modules, this change modifies orderByLatest to be a
helper function that returns an ORDER BY sql string used in many
functions to sort modules.

Updates golang/go#37714

Change-Id: I5c8e5264e90442a960328439dcc445397fc4cdbd
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/248497
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>