internal/fetch/internal/doc: support examples on methods from embedded unexported types
This CL uses the same logic as implemented in CL 249557.

In

  type T1 struct { t2 }
  type t2 int
  func (t2) M()

T1 has method M because it embeds t2, which has M. Classify the example

  func ExampleT1_M

with T1 instead of ignoring it, as is done currently. There is no other
way to provide an example for such a method, since its original type is
unexported.

Continue to ignore examples on methods from embedded types that are
exported, unless in AllMethods mode. Examples for those methods could be
written on the original type.

The change involves removing a check in classifyExamples. The check
isn't necessary to get the above behavior because
reader.collectEmbeddedMethods and sortedFuncs already generate the
appropriate list of methods.

Updates golang/go#40172

Change-Id: Iaeffc1dcfcfca42dc91db1e2527b10286c06fa84
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/249638
Reviewed-by: Jonathan Amsterdam <jba@google.com>