internal: change signature of GetDirectory
The signature of GetDirectory is changed to:

GetDirectory(ctx context.Context, fullPath, modulePath, version string,
    pathID int, fields []internal.FieldSet) (*internal.Directory, error)

The allows the caller to pass in a pathID, if known, and specify
specific fields that they want to fetch, for optimization.

No functional changes to GetDirectory are made in this CL, just an
update to the DatsSource interface, and how it is used in
internal/frontend.

For golang/go#39629

Change-Id: Ic44225a13bf2cf04a62af652dbb830c4d256cb37
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/250799
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>