internal/frontend: use internal.V1Path in fetchVersionDetails
fetchVersionDetails now uses internal.V1Path to compute the V1Path,
instead of accepting v1Path as an argument. This is the only use case
for V1Path by the frontend outside the legacy code, so V1Path no longer
needs to be supported by the new data model structs.

For golang/go#39629

Change-Id: Ia9c44c625f079f190ac212b5085078712d73dd23
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/251417
Reviewed-by: Jonathan Amsterdam <jba@google.com>