internal/{config,log}: configure logging for GKE
Set the monitored resource and log labels so that when we run on GKE,
the logs will show up where we expect them to, under the "Kubernetes
Container" resource in the log viewer.

Clicking on the logs from the GKE console
(e.g. https://pantheon.corp.google.com/kubernetes/workload?project=go-discovery-exp)
will also work.

Also, log to stdout on GKE because logs to stderr have Error severity,
which is distracting.

Change-Id: I3a60800587cfefc6e158e76e4606b932d48ddeae
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/252898
Reviewed-by: Julie Qiu <julie@golang.org>