internal: create GetNestedModules function in DataSource
This change creates a function in the DataSource interface that
retrieves the nested modules as ModuleInfos given a path to a directory.

Note that proxydatasource is included since it is required to fill the
datasource interface but only returns an empty slice since we are not
implementing the functionality there.

This change also adds a scanModuleInfo helper function for scanning
ModuleInfo from the db.

Updates golang/go#38596

Change-Id: I02f25aad2640a0828e58d622d1c8661e6a303b84
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/253742
Reviewed-by: Julie Qiu <julie@golang.org>