internal/godoc: create package for doc rendering
An internal/godoc package is created, which can be used to parse
specific sections of the dochtml in internal/fetch/dochtml.

This package is intended to aid frontend development for the unit page,
and is a temporary solution for prototyping the new designs. It is not
intended for use in production without further testing.

Change-Id: I5d018fa479717e8ed1690312fae1f412296769ba
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/253743
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>