internal: remove use-path-info for latestMinorVersion
This is a final step towards removing the use-path-info experiment.

The use-path-info experiment flag is removed from latestMinorVersion.
This means that the latest badge should be displayed even on directory
pages, which is a different behavior than when the experiment is off.

Tests in frontend.TestServer and TestModulePackageDirectoryResolution
are updated accordingly.

For golang/go#39629

Change-Id: I197cc40924aaeafb80e140a2a14bafd627f372f0
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/255347
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>