internal/fetch: convert ast.Files to and from bytes
Add a pair of functions to encode and decode ast.Files.

We'll use these to store the files in the DB so we
can generate documentation on the frontend.

Originally we stored the full source in a zip file.
That turned out to be too slow.

Then we tried to remove unexported functions and function bodies from
the source before saving it. The problem is that doing so changes the
line numbers, so the source links in the rendered doc point to the
wrong lines. We could perhaps fix that by inserting "//line" compiler
directives in the right places, but it would be tricky.

Instead, we just encode the AST directly, using gob. Gob doesn't
handle cycles, but the fields that are responsible for the cycles
aren't needed for doc generation anyway. It is important to get
the ast.Object pointers right, so we do that.

Change-Id: I79f49b125adfc4e516f0aea6ad91cd5259b3d012
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/257757
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>