internal/godoc: store the set of module packages
In order to render links to packages in the same module,
we need the list of those packages. To keep things simple,
we store the complete set in each Package of the module.

This is wasteful for modules with many packages. Some alternatives:
- Store the list in a new DB column in modules.
- Always read a unit's subdirectories when reading the unit.

Change-Id: Ic2d51444c786b69fcba1717849db86d27d6c7a80
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/258421
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>