internal/godoc,fetch: move doc rendering to godoc package
Write a Render method in the godoc package that does all
doc rendering.

Modify the internal/fetch package to use it.

This required moving the dochtml and internal/doc packages under
internal/godoc.

It's redundant to have the set of module package paths in both
godoc.Package and dochtml.ModuleInfo, but we'll leave that
for now to keep the number of changes to a minimum.

The new Render method is effectively tested by TestFetchModule
in internal/fetch.

Change-Id: I655451386c55c883682eae4277d2d5b2d9c4fc3b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/258423
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>