internal/godoc: remember and check GOOS/GOARCH
Package records the GOOS/GOARCH combination it was created with;
Render checks it.

Someday we'll allow multiple combos, but it requires a bit of
refactoring in internal/fetch, and could conceivably cause us to fail
on packages that we now succeed on (by parsing a file from a build
context that we don't currently get to). So we'll leave it for later.
See golang/go#37232.

Change-Id: I3d5e608fefebf4a9e825d1f8f8a5d1166b7e573b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/258737
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>