content,internal: add source files to unit page
An initial version of the source files section is added to the unit
page. The content for this section should be present, but needs to be
styled in a future CL accordingly to spec.

Change-Id: I398bcd3052048567e5c3a9695427afc3a85a0731
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/259005
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>