internal/frontend: cache crawled pages only briefly
If the User-Agent header on a details request is that of a known
crawler, cache the page very briefly, instead of for 24 hours.

Since much of our traffic is crawlers and their access patterns do not
correspond to real users (in that they're unlikely to revisit a page
soon), it doesn't make sense to cache the pages they hit for a long
time.

We expect this to reduce the large growth in cache size we've seen
recently.

For b/170492779

Change-Id: I81e05148da11e4f6a988a9632e87a324097cceec
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/261821
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>