content: display license message only for doc section
Source files and directories can be displayed even if the unit is not
redistributable, since this is technical information about the unit.

There is no way of knowing whether a README doesn't exist because the
file isn't there, or if it was removed because the unit is not
redistributable.

As a result, the message informing users that a unit is not
redistributable should only be displayed if a documentation section
would be present otherwise.

Change-Id: Ia702a5090a8693da24ccffd5a1be040072fe188e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/262100
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>