internal/godoc/codec: fast encoding of ASTs
This is the first of a series of CLs that implements a new way to turn
an AST from the go/ast package into a sequence of bytes.

The new encoding was motivated by the slowness of decoding an AST
using gob. Since that happens on the serving path, it is important to
make it fast. In some of our measurements it was a significant
fraction, 20% or so, of the total serving time.

(The encoding happens on the worker, where it doesn't affect the user
experience, so we are less concerned with fast encoding than fast
decoding.)

Another problem with gob was that it doesn't handle sharing, and
go/ast structures have a lot of sharing, including cycles. For
example, an ast.Ident inside an ast.FuncDecl points to an ast.Object
which points back to the FuncDecl. Preserving that sharing is
important for generating documentation (the ultimate goal of saving
the AST). Currently we have some tricky code in
internal/godoc/encoding.go to undo that sharing before gob-encoding
and restore it after. Writing a new encoder provides the opportunity
to handle sharing there, where it is simpler and more general.

The encoder is presented bottom-up. This first CL defines the Encoder
and Decoder types and the methods that perform low-level I/O of bytes,
strings and unsigned integers. Later CLs will implement the encoding
of Go types using these building blocks.

Change-Id: I7606397ecf249c02b375b2882382975a20e8491a
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264117
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>