internal/godoc/codec: encoding lists
The nValues byte code is used to encode a list whose elements might
differ in size.  After the nValues byte is a uint with the size,
followed by the encodings of the values.  The StartList methods on
Encoder and Decoder provide support.

There is a special case when a list encodes a Go slice: the slice
might be nil. The Decoder looks for a 0 byte to indicate that.  The
corresponding encoding happens when code is generated for Go types, in
a later CL.

Change-Id: Iab5b11ccc7e51a9a759cdbaefd731d9010c98f65
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264122
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>