internal/godoc/codec: encoding registered types
This CL starts the work for encoding an arbitrary type. EncodeAny
accepts an interface{} and tries to encode it. To do so, the value's
type has to be registered with an encoding and decoding function.
The encoder stores a number identifying the type along with its
encoded value.

In the next CL we'll see how the Decoder uses the Encoder's
type information along with the registry to decode types.

Change-Id: I60e8766adfaff9622e22d474280e55b401460a62
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264123
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>