internal/godoc/codec: decode arbitrary values
This CL adds DecodeAny to complete EncodeAny from the previous CL.

The Decoder gets the list of encoded types from the beginning of the
encoded byte slice, and uses the type names to look up the typeInfos
in the registry. It can then use the type numbers written by the
Encoder to find the right function to decode the value.

Change-Id: Ia5376eed94cf68f0d7068e89d58963e00cc83cb7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264124
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>