internal: return 404 on DecodePackage error
When an error is returned by DecodePackage in fetchMainDetails, the
issue is generally that we need to reprocess. Show the user the 404 page
that enables them to do that instead of a 500.

Change-Id: Ia747b6b2e0d947474e8a1b5943573773d68a41ba
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/263981
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>