internal/godoc/codec: encoding structs
This CL contains both support (in codec.go) and generated code (in
generate.go) for encoding structs.

Structs are encoded using the `start` and `end` codes, with a list of
(field, value) pairs between them. Fields are represented by a number,
currently the related to index in the list of fields exposed by the
reflect package. Using a number rather than the field name takes up
less space, and lets us write an efficient switch statement to decode
the fields. But it can cause a problem if a new field is added or
existing fields are rearranged, which we will deal with in a later CL.

If the decoder doesn't know a field number, then a new field was added
since the decoder was built. In that case we abort the decoding. In
the next CL, we will instead skip over the unknown field.

Change-Id: I51b55e05e71e38a293e6aaa2613571a86540e6fc
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264261
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>