internal/godoc/codec: skip unknown fields
Instead of failing if we see an unknown field, skip the value and keep
going. That way an old decoder can read a struct encoded with a newer
encoder. The encoding is designed so that a value can be skipped
without knowing its type or anything else about it.

Change-Id: I4f776de01f637cd11a488c78f0563cc2ac743405
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264597
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>