internal/godoc/codec: handle sharing
One of the motivations for this work was handling the pointer sharing
that occurs in ASTs. One example is the list of *ast.Idents in
ast.File.Unresolved, which contains the same pointers as appear in the
AST itself. Another is the cycle that occurs when an ast.Ident inside
a Decl holds an *ast.Object that points to the Dec.

gob doesn't handle sharing. It fails on cycles, and on non-cyclical
sharing like the first example, it will succeed but will encode two
copies of the shared value, instead of a single one pointed to from
both locations.

This CL implements proper treatment of sharing. When the encoder
encounters a struct pointer, it first checks whether it has seen it
before. If it hasn't, it encodes the struct as usual and remembers it.
If it has, it instead outputs a reference to the position of the
struct in the encoded output (the first struct gets 0, the second 1,
and so on).

Whenever the decoder sees a struct pointer, it remembers it in a
slice. When decoding a struct, if it sees a ref to N, it returns the
pointer at index N in the slice.

Change-Id: I7a787b25c9000cc4fa507a5804ec684dc4d7b9e7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264598
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>