internal/godoc/codec: preserve field numbers across changes
Using field numbers in the encoding instead of names
means that the numbers can change if the struct changes.
This CL addresses that problem.

Here is how the problem can happen:

1. An encoder and decoder for a struct are created.
2. A new struct field is added in the middle, or the
   existing fields are rearranged.
3. A new encoder and decoder are built.
4. The old decoder from step 1 attempts to decode
   a struct encoded with the new encoder.

The problem is not just that there is a new field--the decoder
can handle that, by skipping its value. The problem is that
the field numbers of existing fields changed in step 2.
For example, if we start with

  type S struct { A, B int }

and then decide to swap the fields:

  type S struct { B, A int }

then the old decoder in step 5, still using the field numbers 0 for A
and 1 for B, will put the value of A into B and vice versa.

We address this by remembering the field numbers in the generated
file, in a comment. When re-generating the file, we read the comment
and preserve the numbers.

Change-Id: I097b76d71773e358be00308d789dcf4aa4e54230
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/264599
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>