internal/testing/sample,etc.: add package source
Add encoded source to the sample Documentation value, and render it to
get the documentation HTML.  Use the package in
internal/godoc/testdata.

I had hoped to use the godoc package in an init function to create the
needed values, but that caused an import cycle between sample and the
godoc test package. So instead, a small program invoked with `go
generate` uses godoc to generate the values and embed them in a Go
file.

This change required a few other changes:

internal/postgres:  Always insert source when comparing Documentation.

internal/frontend: Change the strings we look for in doc to match the
one in the sample package.

internal/frontend: Remove the check on HTML ids. This was added when I
was doing safehtml, when I was being strict about HTML ids. The
safehtml package wants ids to conform to some strict rules to avoid
certain attacks, and I tried to follow those rules. But in the end I
had to be more forgiving, or I would have broken anchor tags to Go
symbols, like pkg.go.dev/fmt#Println. (safehtml does not consider
"Println" a safe identifier unless it is a constant string.)

The checkIDs function was added somewhere in the middle of that
process, and it worked until now because the value of
sample.DocumentationHTML was an ordinary text string with no ids, not
real HTML doc.  It was actually only checking the ids for license
files, and those do contain a hyphen. I don't think its loss is
important.

Change-Id: I4019c56b216aaf1a4e92a6167c2af1eef4cfbf89
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/265297
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>