internal/postgres: skip package versions with no source
In GetVersionForPath, if the path is for a package and there is no
source in the DB, don't return the version.

This avoids an error when the user clicks on a version in the versions
list for a package and gets a 404. For example,
https://beta.pkg.go.dev/github.com/aws/aws-sdk-go@v1.34.0/service/xray/xrayiface.

This is a workaround for our current inability to process some large
modules.

Change-Id: Iab982bde9ec8916b0e600809e065a3473f1cb6dc
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/265897
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>