internal/godoc,content/static/html: move templates to files
Move the doc rendering templates from hard-coded
strings to files under content/static/html.

Also, parse the templates once, instead of on each call to Render.

Template loading is still under the control of the
internal/godoc/dochtml package, because the templates have to be
parsed with a certain grouping and with some functions defined, and it
would be awkward if callers had to know all that.

An alternative design would pass the directory containing the template
files to Render, but that would require a lot of plumbing.

This is the first of several improvements to internal/godoc.
We also plan to remove godoc.Parse.

Change-Id: I7182dd4ea9bc7189b1144295ad193af30ef9ecc4
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/266597
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>