internal/postgres: determine path/unit_id column dynamically
Three tables still have a path_id column that is a foreign key for the
"units" (previously "paths") table. It is not possible to rename those
columns while keeping the code working: we would have to create a new
column, change the code, then delete the old column, all the while
keeping the columns in sync.

This CL suggests a simpler approach: the code figures out the name of
the column on its own.  Then we can simply do one migration to rename
the all the path_id to unit_id columns, and the code will continue to
work.

The code becomes uglier, but it is temporary: after we're done, we'll
put it back the way it was, with literal column names.

Change-Id: Ia4f621e1c0112d378cf63811311a430982c1cb5f
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/267238
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>