migrations: add indexes to paths table
A UNIQUE index is add to paths.path, since values in that column
are unique.

An index is added to paths(path, id), which should improve performance
for reverse lookups from path to id, since the index will be sorted by
path not id.

Change-Id: I51df32ce2e85835f89defc6b9e7eeee97112a7be
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/267457
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Trust: Julie Qiu <julie@golang.org>