internal/frontend: update fetch 500 logs and error page
When a frontend fetch request returns a 500, an internal server error is
now logged instead of errModuleDoesnotExist.

Additionally, the user will be presented with a 404 page with
instructions on how to fetch the path via the proxy. The option for a
frontend fetch when a request previously returned a 500 is removed.

Change-Id: I68548b45c0702424a153382eb1dc8de5009704ae
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/267698
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>