internal/frontend: redirect if tab is invalid
When a tab is invalid for a pageType, redirect the user to the main
page. At the moment, users can visit an empty imports and imported by
tab when the pageType is not a package.

Change-Id: I43e15d3d124ed5edb735b48996b6de219f876f49
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/267683
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>