internal/worker: don't delete module if load-shedding
If the worker decides to shed an incoming module, it should not take
the time to delete it. For large modules that can be a while
(e.g. over three seconds for
github.com/nhanvpt102/azure-sdk-for-go@v32.5.0+incompatible) and adds
load to the DB.

Change-Id: I577c6c6d7c672829f4828c49e013a32af21237ba
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/268398
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>