internal/frontend: update not found page on DecodePackage erorr
When a DecodePackage error occurs, the path will still be marked as a 200
in the version_map table. Return instructions on how to fetch the path instead
of a the request button, otherwise this will result in a loop.

Change-Id: I7074515ddddb8a2f727fab4b0c3b2099ca94dc03
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/268840
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>