internal/frontend: approximate imported-by count on unit page
It's annoying that the imported-by count on the unit page disagrees
with the value on the imported-by tab. But it's too expensive to get a
list of all the importers just to compute the number.

Instead, get an exact count for small values by listing the importers
with a small limit. If there are more than the limit, then use
search_documents.imported_by_count, as we did previously, but make it
lear that it's an approximation.

Change-Id: I95ccc4aa30be891411b1d415e9e2f68c03faffd9
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/268860
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>