internal/frontend: don't precheck 404s
Previously when a user received a 404, we checked all the candidate
module paths to determine if the path was recently fetched. If so, the
user is not presented with a Request button. This was done to create a
nicer user experience, but it also can result in a lot of queries when
we are being scraped. This check is removed to reduce queries in that
case.

Change-Id: If8917b5ddf6374fe17a742bd8d255fc6ced89e94
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/270266
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>